+++ src/icommon.h
@@ -3,7 +3,7 @@
 /* icommon.h - common interactive routines            */
 /*                                                    */
 /******************************************************/
-/* Copyright 2015,2016,2017,2019 Pierre Abbat.
+/* Copyright 2015,2016,2017,2019,2022 Pierre Abbat.
  * This file is part of Bezitopo.
  *
  * Bezitopo is free software: you can redistribute it and/or modify
@@ -48,6 +48,6 @@ extern bool subcont; // continue flag within commands
 
 void setfoot_i(std::string args);
 void setlengthunit_i(std::string args);
-arangle parsearangle(std::string angstr,int unitp);
+arangle parsearangle(std::string angstr,int64_t unitp);
 xy parsexy(std::string xystr);
 void subexit(std::string args);
